server/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
minio==7.1.15
orjson==3.8.9
passlib==1.7.4
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.22.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.25.11
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0
Werkzeug==2.2.3
